Move previous edit sections into their own views
