+++ Mastodon/In Progress New Layout and Datamodel/Profile/ProfileEditingView.swift
@@ -284,7 +284,7 @@ struct ORIGINALProfileEditingView: View {
                     Divider()
                     
                     if editingViewModel.showTabDisplayPreferences {
-                        DisplayPreferencesEditor()
+                        ProfileTabSettingsEditor()
                             .padding(.horizontal)
                             .padding(.vertical, doublePadding)
                         
@@ -804,214 +804,6 @@ struct ProfileSectionHeader: View {
             Text(title)
                 .font(.title3)
                 .fontWeight(.semibold)
-        }
-    }
-}
-
-struct CustomProfileFieldsEditor: View {
-    @Environment(ProfileEditingViewModel.self) var editingViewModel
-    @State private var draggingField: Mastodon.Entity.Field?
-    
-    var body: some View {
-        VStack(alignment: .leading) {
-            ProfileSectionHeader(section: .customFields)
-            infoButton(.customFields)
-            SubsectionHeading(title: nil, subtitle: "Add your pronouns, external links, or anything else you’d like to share.") // TODO: needs L10n
-            if let customFields = editingViewModel.customFields, !customFields.isEmpty {
-                customFieldsList(customFields)
-                Divider()
-            }
-            addFieldButton
-            if editingViewModel.showVerifiedLinkTip {
-                verifiedLinksTipBox
-            }
-        }
-    }
-    
-    @ViewBuilder var addFieldButton: some View {
-        Button() {
-            withAnimation {
-                editingViewModel.beginEditingField(.create)
-            }
-        } label: {
-            HStack(spacing: tinySpacing) {
-                Image(systemName: "plus")
-                Text("Add a field") // TODO: needs L10n
-            }
-            .font(.subheadline)
-            .padding(.vertical, tinySpacing)
-            .padding(.horizontal, standardPadding)
-            .background() {
-                Capsule()
-                    .fill(.quinary)
-            }
-        }
-    }
-    
-    @ViewBuilder var verifiedLinksTipBox: some View {
-        HStack(alignment: .top) {
-            Image(systemName: "checkmark.seal")
-                .font(.subheadline)
-                .padding(tinySpacing)
-                .background() {
-                    Circle()
-                        .fill(brandBackgroundColor)
-                }
-            
-            VStack(alignment: .leading) {
-                Spacer()
-                    .frame(height: tinySpacing)
-                Text("Tip: Adding verified links")
-                    .fontWeight(.semibold)
-                Text("You can easily add credibility to your Mastodon account by verifying links to any websites you own.")
-            }
-            .font(.subheadline)
-            
-            Button() {
-                withAnimation {
-                    editingViewModel.showVerifiedLinkTip = false
-                }
-            } label: {
-                Image(systemName: "xmark")
-            }
-        }
-        .padding()
-        .background() {
-            RoundedRectangle(cornerRadius: CornerRadius.extraLarge)
-                .fill(brandBackgroundColor)
-        }
-    }
-    
-    @ViewBuilder func customFieldsList(_ customFields: [Mastodon.Entity.Field]) -> some View {
-        VStack {
-            ForEach(customFields, id: \.self) { field in
-                customFieldRow(field, isDraggable: false)
-                    .onTapGesture {
-                        withAnimation {
-                            editingViewModel.beginEditingField(.edit(field))
-                        }
-                    }
-            }
-        }
-    }
-    
-    @ViewBuilder func customFieldRow(_ field: Mastodon.Entity.Field, isDraggable: Bool, isDragging: Bool) -> some View {
-        if isDragging {
-            customFieldRow(field, isDraggable: isDraggable)
-                .hidden()
-        } else {
-            customFieldRow(field, isDraggable: isDraggable)
-        }
-    }
-    
-    @ViewBuilder func customFieldRow(_ field: Mastodon.Entity.Field, isDraggable: Bool) -> some View {
-        HStack {
-            if isDraggable {
-                Image(systemName: "line.3.horizontal")
-            }
-            
-            VStack(alignment: .leading) {
-                Text(field.name)
-                    .fixedSize(horizontal: false, vertical: true)
-                MastodonContentView.customProfileField(html: field.value, emojis: editingViewModel.emojis, bold: true, lineLimit: 1)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            .font(.footnote)
-            
-            Spacer()
-            
-            Button() {
-                withAnimation {
-                    editingViewModel.requestDeleteCustomField(field)
-                }
-            } label: {
-                Image(systemName: "trash")
-                    .font(.subheadline)
-                    .fontWeight(.semibold)
-                    .foregroundColor(.red)
-                    .padding(standardPadding)
-                    .background() {
-                        Circle()
-                            .stroke(.quaternary)
-                    }
-            }
-        }
-        .padding()
-    }
-}
-
-struct DisplayPreferencesEditor: View {
-    @Environment(ProfileViewModel.self) var profileViewModel
-    @Environment(ProfileEditingViewModel.self) var editingViewModel
-    
-    var body: some View {
-        VStack(alignment: .leading, spacing: doublePadding) {
-            VStack(alignment: .leading) {
-                ProfileSectionHeader(section: .displayPreferences)
-                infoButton(.displayPreferences)
-            }
-            
-            // SHOW/HIDE MEDIA TAB
-            VStack(alignment: .leading) {
-                SubsectionHeading(title: "’Media’ tab settings", subtitle: "‘Media’ is an optional tab that shows your posts containing images or videos.")
-                RadioButtonArray(items: [(ProfileEditingViewModel.MediaTabVisibilitySetting.showMediaTab.rawValue, "Show ‘Media’ tab"), (ProfileEditingViewModel.MediaTabVisibilitySetting.hideMediaTab.rawValue, "Hide ‘Media’ tab")], selectedItem: Binding<Int>(
-                    get: { editingViewModel.mediaTabVisibilitySetting.rawValue },
-                    set: { newValue in
-                        guard newValue != editingViewModel.mediaTabVisibilitySetting.rawValue else { return }
-                        editingViewModel.setMediaTabVisibilitySetting(.init(rawValue: newValue) ?? .showMediaTab)
-                        Task {
-                            try await profileViewModel.commitTabSettingsChanges()
-                        }
-                    }
-                ))
-            }
-            
-            // INCLUDE REPLIES ON MEDIA TAB
-            VStack(alignment: .leading) {
-                SubsectionHeading(title: "Include replies on ’Media’ tab?", subtitle: nil)
-                RadioButtonArray(items: [(ProfileEditingViewModel.MediaTabRepliesSetting.showDirectPostsOnly.rawValue, "Only show my posts"), (ProfileEditingViewModel.MediaTabRepliesSetting.includeMyRepliesToOthers.rawValue, "Show my posts and replies to other people's posts")], selectedItem: Binding<Int>(
-                    get: { editingViewModel.mediaTabRepliesSetting.rawValue },
-                    set: { newValue in
-                        guard newValue != editingViewModel.mediaTabRepliesSetting.rawValue else { return }
-                        editingViewModel.setMediaTabRepliesSetting(.init(rawValue: newValue) ?? .showDirectPostsOnly)
-                        Task {
-                            try await profileViewModel.commitTabSettingsChanges()
-                        }
-                    }
-                ))
-            }
-            
-            // FEATURED TAB SETTINGS
-            VStack(alignment: .leading) {
-                SubsectionHeading(title: "’Featured’ tab settings", subtitle: "’Featured’ is an optional tab where you can showcase other accounts and collections.")
-                RadioButtonArray(items: [(ProfileEditingViewModel.FeaturedTabVisibilitySetting.showFeaturedTab.rawValue, "Show ’Featured’ tab"), (ProfileEditingViewModel.FeaturedTabVisibilitySetting.hideFeaturedTab.rawValue, "Hide ’Featured’ tab")], selectedItem: Binding<Int>(
-                    get: { editingViewModel.featuredTabVisibilitySetting.rawValue },
-                    set: { newValue in
-                        guard newValue != editingViewModel.featuredTabVisibilitySetting.rawValue else { return }
-                        editingViewModel.setFeaturedTabVisibilitySetting(.init(rawValue: newValue) ?? .showFeaturedTab)
-                        Task {
-                            try await profileViewModel.commitTabSettingsChanges()
-                        }
-                    }
-                ))
-            }
-            
-            Divider()
-            
-            // FEATURED HASHTAGS
-            
-            HStack {
-                SubsectionHeading(title: "Featured hashtags", subtitle: "Help others identify, and have quick access to, your favorite topics")
-                HStack {
-                    Text("Manage")
-                    Image(systemName: "chevron.forward")
-                }
-                .font(.subheadline)
-                .fontWeight(.semibold)
-            }
-            .onTapGesture {
-                // TODO: navigate
-            }
         }
     }
 }
